second commit

+++ backend/routes/posts.js
@@ -42,7 +42,8 @@ router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, r
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
     
     post.save().then(createdPost => {
@@ -98,18 +99,26 @@ router.put(
         imagePath: imagePath
       });
       console.log(post);
-      Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({ message: "Update successful!" });
+      Post.updateOne({ _id: req.params.id , creator: req.userData.userId }, post).then(result => {
+          if(result.nModified > 0){
+            res.status(200).json({ message: "Update successful!" });
+          }
+          else{
+            res.status(401).json({ message: "Access Denied!" });
+          }
+        
       });
     }
   );
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        console.log(result);
-        res.status(200).json({
-            message: 'Post deleted successfully!'
-        })
+    Post.deleteOne({ _id: req.params.id , creator: req.userData.userId }).then(result => {
+        if(result.n > 0){
+            res.status(200).json({ message: "Deletion successful!" });
+          }
+          else{
+            res.status(401).json({ message: "Access Denied!" });
+          }
     })
         .catch(err => {
             console.log('error: ', err);
